+++ src/Native/LibTorchSharp/THSTensor.h
@@ -1026,6 +1026,10 @@ EXPORT_API(Tensor) THSTensor_relu6_(const Tensor tensor);
 
 EXPORT_API(Tensor) THSTensor_repeat(const Tensor tensor, const int64_t* sizes, const int length);
 
+EXPORT_API(Tensor) THSTensor_repeat_interleave(const Tensor tensor, const Tensor repeats, const int64_t dim, const int64_t output_size);
+
+EXPORT_API(Tensor) THSTensor_repeat_interleave_int64(const Tensor tensor, const int64_t repeats, const int64_t dim, const int64_t output_size);
+
 EXPORT_API(int) THSTensor_requires_grad(const Tensor tensor);
 
 EXPORT_API(Tensor) THSTensor_reshape(const Tensor tensor, const int64_t* shape, const int length);
@@ -1175,6 +1179,8 @@ EXPORT_API(Tensor) THSTensor_dstack(const Tensor* tensor, const int length);
 EXPORT_API(Tensor) THSTensor_column_stack(const Tensor* tensor, const int length);
 
 EXPORT_API(Tensor) THSTensor_row_stack(const Tensor* tensor, const int length);
+
+EXPORT_API(void) THSTensor_meshgrid(const Tensor* tensors, const int64_t length, const char* indexing, Tensor* (*allocator)(size_t length));
 
 EXPORT_API(int64_t) THSTensor_stride(const Tensor tensor, const int64_t dim);