+++ src/TorchSharp/Tensor/Tensor.cs
@@ -2030,7 +2030,7 @@ namespace TorchSharp
             static extern IntPtr THSTensor_all(IntPtr tensor);
 
             /// <summary>
-            ///
+            /// Tests if all elements in input evaluate to true.
             /// </summary>
 
             public Tensor all()
@@ -2045,10 +2045,10 @@ namespace TorchSharp
             static extern IntPtr THSTensor_all_along_dimension(IntPtr tensor, long dimension, bool keep_dim);
 
             /// <summary>
-            ///
+            /// Tests if all elements in input evaluate to true.
             /// </summary>
-            /// <param name="dimension"></param>
-            /// <param name="keepDim"></param>
+            /// <param name="dimension">The dimension to reduce</param>
+            /// <param name="keepDim">Keep the dimension to reduce</param>
 
             public Tensor all(long dimension, bool keepDim = false)
             {
@@ -2170,7 +2170,7 @@ namespace TorchSharp
             static extern IntPtr THSTensor_any(IntPtr tensor);
 
             /// <summary>
-            ///
+            /// Tests if any element in input evaluate to true.
             /// </summary>
 
             public Tensor any()
@@ -2185,10 +2185,10 @@ namespace TorchSharp
             static extern IntPtr THSTensor_any_along_dimension(IntPtr tensor, long dimension, bool keep_dim);
 
             /// <summary>
-            ///
+            /// Tests if any element in input evaluate to true.
             /// </summary>
-            /// <param name="dimension"></param>
-            /// <param name="keepDim"></param>
+            /// <param name="dimension">The dimension to reduce</param>
+            /// <param name="keepDim">Keep the dimension to reduce</param>
 
             public Tensor any(long dimension, bool keepDim = false)
             {
@@ -5149,6 +5149,30 @@ namespace TorchSharp
             }
 
             [DllImport("LibTorchSharp")]
+            extern static IntPtr THSTensor_repeat_interleave(IntPtr tensor, IntPtr repeats, long dim, long output_size);
+
+            public Tensor repeat_interleave(Tensor repeats, long? dim = null, long? output_size = null)
+            {
+                long _dim = dim ?? long.MinValue;
+                long _output_size = output_size ?? long.MinValue;
+                var res = THSTensor_repeat_interleave(Handle, repeats.Handle, _dim, _output_size);
+                if (res == IntPtr.Zero) { torch.CheckForErrors(); }
+                return new Tensor(res);
+            }
+
+            [DllImport("LibTorchSharp")]
+            extern static IntPtr THSTensor_repeat_interleave_int64(IntPtr tensor, long repeats, long dim, long output_size);
+
+            public Tensor repeat_interleave(long repeats, long? dim = null, long? output_size = null)
+            {
+                long _dim = dim ?? long.MinValue;
+                long _output_size = output_size ?? long.MinValue;
+                var res = THSTensor_repeat_interleave_int64(Handle, repeats, _dim, _output_size);
+                if (res == IntPtr.Zero) { torch.CheckForErrors(); }
+                return new Tensor(res);
+            }
+
+            [DllImport("LibTorchSharp")]
             extern static IntPtr THSTensor_broadcast_to(IntPtr tensor, IntPtr psizes, int length);
 
             /// <summary>
@@ -6177,9 +6201,10 @@ namespace TorchSharp
             extern static IntPtr THSTensor_nonzero(IntPtr tensor);
 
             /// <summary>
-            ///
+            /// Returns a tensor containing the indices of all non-zero elements of input.
+            /// Each row in the result contains the indices of a non-zero element in input.
+            /// The result is sorted lexicographically, with the last index changing the fastest (C-style).
             /// </summary>
-
             public Tensor nonzero()
             {
                 var res = THSTensor_nonzero(Handle);